typescript types cleanup
