+++ npm/src/typings.ts
@@ -1,6 +1,7 @@
 import { type JWK } from 'jose';
 
 export * from './ee/federated-saml/types';
+export * from './directory-sync/types';
 
 interface SSOConnection {
   defaultRedirectUrl: string;
@@ -415,274 +416,11 @@ export interface ISPSAMLConfig {
   toXMLMetadata(): Promise<string>;
 }
 
-export type DirectorySyncEventType =
-  | 'user.created'
-  | 'user.updated'
-  | 'user.deleted'
-  | 'group.created'
-  | 'group.updated'
-  | 'group.deleted'
-  | 'group.user_added'
-  | 'group.user_removed';
-
-export interface Base {
-  store(type: 'groups' | 'members' | 'users'): Storable;
-  setTenant(tenant: string): this;
-  setProduct(product: string): this;
-  setTenantAndProduct(tenant: string, product: string): this;
-  with(tenant: string, product: string): this;
-  createId(): string;
-}
-
-export interface Users extends Base {
-  list({
-    pageOffset,
-    pageLimit,
-  }: {
-    pageOffset?: number;
-    pageLimit?: number;
-  }): Promise<{ data: User[] | null; error: ApiError | null }>;
-  get(id: string): Promise<{ data: User | null; error: ApiError | null }>;
-  search(userName: string): Promise<{ data: User[] | null; error: ApiError | null }>;
-  delete(id: string): Promise<{ data: null; error: ApiError | null }>;
-  clear(): Promise<void>;
-  create(param: {
-    first_name: string;
-    last_name: string;
-    email: string;
-    active: boolean;
-    raw: any;
-  }): Promise<{ data: User | null; error: ApiError | null }>;
-  update(
-    id: string,
-    param: {
-      first_name: string;
-      last_name: string;
-      email: string;
-      active: boolean;
-      raw: object;
-    }
-  ): Promise<{ data: User | null; error: ApiError | null }>;
-}
-
-export interface Groups extends Base {
-  create(param: { name: string; raw: any }): Promise<{ data: Group | null; error: ApiError | null }>;
-  removeAllUsers(groupId: string): Promise<void>;
-  list({
-    pageOffset,
-    pageLimit,
-  }: {
-    pageOffset?: number;
-    pageLimit?: number;
-  }): Promise<{ data: Group[] | null; error: ApiError | null }>;
-  get(id: string): Promise<{ data: Group | null; error: ApiError | null }>;
-  getAllUsers(groupId: string): Promise<{ user_id: string }[]>;
-  delete(id: string): Promise<{ data: null; error: ApiError | null }>;
-  addUserToGroup(groupId: string, userId: string): Promise<void>;
-  isUserInGroup(groupId: string, userId: string): Promise<boolean>;
-  removeUserFromGroup(groupId: string, userId: string): Promise<void>;
-  search(displayName: string): Promise<{ data: Group[] | null; error: ApiError | null }>;
-  update(
-    id: string,
-    param: {
-      name: string;
-      raw: any;
-    }
-  ): Promise<{ data: Group | null; error: ApiError | null }>;
-}
-
-export type User = {
-  id: string;
-  email: string;
-  first_name: string;
-  last_name: string;
-  active: boolean;
-  raw?: any;
-};
-
-export type Group = {
-  id: string;
-  name: string;
-  raw?: any;
-};
-
-export enum DirectorySyncProviders {
-  'azure-scim-v2' = 'Azure SCIM v2.0',
-  'onelogin-scim-v2' = 'OneLogin SCIM v2.0',
-  'okta-scim-v2' = 'Okta SCIM v2.0',
-  'jumpcloud-scim-v2' = 'JumpCloud v2.0',
-  'generic-scim-v2' = 'SCIM Generic v2.0',
-}
-
-export type DirectoryType = keyof typeof DirectorySyncProviders;
-
-export type HTTPMethod = 'POST' | 'PUT' | 'DELETE' | 'GET' | 'PATCH';
-
-export type Directory = {
-  id: string;
-  name: string;
-  tenant: string;
-  product: string;
-  type: DirectoryType;
-  log_webhook_events: boolean;
-  scim: {
-    path: string;
-    endpoint?: string;
-    secret: string;
-  };
-  webhook: {
-    endpoint: string;
-    secret: string;
-  };
-};
-
-export type DirectorySyncGroupMember = { value: string; email?: string };
-
-export interface DirectoryConfig {
-  create({
-    name,
-    tenant,
-    product,
-    webhook_url,
-    webhook_secret,
-    type,
-  }: {
-    name?: string;
-    tenant: string;
-    product: string;
-    webhook_url?: string;
-    webhook_secret?: string;
-    type?: DirectoryType;
-  }): Promise<{ data: Directory | null; error: ApiError | null }>;
-  update(
-    id: string,
-    param: Omit<Partial<Directory>, 'id' | 'tenant' | 'prodct' | 'scim'>
-  ): Promise<{ data: Directory | null; error: ApiError | null }>;
-  get(id: string): Promise<{ data: Directory | null; error: ApiError | null }>;
-  getByTenantAndProduct(
-    tenant: string,
-    product: string
-  ): Promise<{ data: Directory[] | null; error: ApiError | null }>;
-  list({
-    pageOffset,
-    pageLimit,
-  }: {
-    pageOffset?: number;
-    pageLimit?: number;
-  }): Promise<{ data: Directory[] | null; error: ApiError | null }>;
-  delete(id: string): Promise<void>;
-}
-
-export interface IDirectoryUsers {
-  create(directory: Directory, body: any): Promise<DirectorySyncResponse>;
-  get(user: User): Promise<DirectorySyncResponse>;
-  update(directory: Directory, user: User, body: any): Promise<DirectorySyncResponse>;
-  patch(directory: Directory, user: User, body: any): Promise<DirectorySyncResponse>;
-  delete(directory: Directory, user: User, active: boolean): Promise<DirectorySyncResponse>;
-  getAll(queryParams: { count: number; startIndex: number; filter?: string }): Promise<DirectorySyncResponse>;
-  handleRequest(request: DirectorySyncRequest, eventCallback?: EventCallback): Promise<DirectorySyncResponse>;
-}
-
-export interface IDirectoryGroups {
-  create(directory: Directory, body: any): Promise<DirectorySyncResponse>;
-  get(group: Group): Promise<DirectorySyncResponse>;
-  updateDisplayName(directory: Directory, group: Group, body: any): Promise<Group>;
-  delete(directory: Directory, group: Group): Promise<DirectorySyncResponse>;
-  getAll(queryParams: { filter?: string }): Promise<DirectorySyncResponse>;
-  addGroupMembers(
-    directory: Directory,
-    group: Group,
-    members: DirectorySyncGroupMember[] | undefined,
-    sendWebhookEvent: boolean
-  ): Promise<void>;
-  removeGroupMembers(
-    directory: Directory,
-    group: Group,
-    members: DirectorySyncGroupMember[],
-    sendWebhookEvent: boolean
-  ): Promise<void>;
-  addOrRemoveGroupMembers(
-    directory: Directory,
-    group: Group,
-    members: DirectorySyncGroupMember[]
-  ): Promise<void>;
-  update(directory: Directory, group: Group, body: any): Promise<DirectorySyncResponse>;
-  patch(directory: Directory, group: Group, body: any): Promise<DirectorySyncResponse>;
-  handleRequest(request: DirectorySyncRequest, eventCallback?: EventCallback): Promise<DirectorySyncResponse>;
-}
-
-export interface IWebhookEventsLogger extends Base {
-  log(directory: Directory, event: DirectorySyncEvent): Promise<WebhookEventLog>;
-  getAll({ pageOffset, pageLimit }: { pageOffset?: number; pageLimit?: number }): Promise<WebhookEventLog[]>;
-  get(id: string): Promise<WebhookEventLog>;
-  clear(): Promise<void>;
-  delete(id: string): Promise<void>;
-  updateStatus(log: WebhookEventLog, statusCode: number): Promise<WebhookEventLog>;
-}
-
-export type DirectorySyncResponse = {
-  status: number;
-  data?: any;
-};
-
-export interface DirectorySyncRequestHandler {
-  handle(request: DirectorySyncRequest, callback?: EventCallback): Promise<DirectorySyncResponse>;
-}
-
-export interface Events {
-  handle(event: DirectorySyncEvent): Promise<void>;
-}
-
-export interface DirectorySyncRequest {
-  method: HTTPMethod;
-  body: any | undefined;
-  directoryId: Directory['id'];
-  resourceType: 'users' | 'groups';
-  resourceId: string | undefined;
-  apiSecret: string | null;
-  query: {
-    count?: number;
-    startIndex?: number;
-    filter?: string;
-  };
-}
-
-export type IDirectorySyncController = {
-  requests: DirectorySyncRequestHandler;
-  directories: DirectoryConfig;
-  groups: Groups;
-  users: Users;
-  events: { callback: EventCallback };
-  webhookLogs: IWebhookEventsLogger;
-  providers: () => {
-    [K in string]: string;
-  };
-};
-
 export interface ApiError {
   message: string;
   code: number;
 }
 
-export interface DirectorySyncEvent {
-  directory_id: Directory['id'];
-  event: DirectorySyncEventType;
-  data: User | Group | (User & { group: Group });
-  tenant: string;
-  product: string;
-}
-
-export interface EventCallback {
-  (event: DirectorySyncEvent): Promise<void>;
-}
-
-export interface WebhookEventLog extends DirectorySyncEvent {
-  id: string;
-  webhook_endpoint: string;
-  created_at: Date;
-  status_code?: number;
-  delivered?: boolean;
-}
 export type SetupLinkCreatePayload = {
   tenant: string;
   product: string;
